feat: Changing register and login page with click Add the function to chenge pages when the blue text is clicked
